converted NextVersionTask to simple function
