+++ build.gradle.kts
@@ -7,7 +7,7 @@ plugins {
     id("pl.droidsonroids.jacoco.testkit") version "1.0.9"
     `maven-publish`  // TODO REMOVE, just for local tests
 }
-version="1.0.3"  // TODO REMOVE, just for local tests
+version="1.0.4"  // TODO REMOVE, just for local tests
 description = "gradle semantic versioning plugin"
 group = "ch.fuzzle.gradle.semver"
